fix expr constant and improve slot extractor
