+++ fe/fe-core/src/main/java/org/apache/doris/nereids/analyzer/UnboundSlot.java
@@ -52,11 +52,6 @@ public class UnboundSlot extends Slot implements Unbound {
     }
 
     @Override
-    public boolean isConstant() {
-        return false;
-    }
-
-    @Override
     public String toSql() {
         return nameParts.stream().map(Utils::quoteIfNeeded).reduce((left, right) -> left + "." + right).orElse("");
     }
